Week 5/Blog_API ___App/requirements.txt
Django>=5.2
djangorestframework
djangorestframework-simplejwt  # Optional if using JWT in future
django-filter
